Change Docker Volume Mount
The current mount does not work. The `$GITHUB_WORKSPACE` environment
variable is not interpolated in the `ci.yml` (at least not for the
`volume` option of the container).

According to the "internet" the workspace is mounted anyway to
"/home/workspace" in the container itself.

So I changed the "workdir" of the container to "$GITHUB_WORKSPACE" and
hope this will fix the issue.