Remove Base32Kit Target
In agreement with Fabian I refactored the code so that there is only one
target left. For now we keep the name "Base64Kit". Changing the name
will be part of a later pull request.

To be able to move my code into the Base64Kit folders I had to do some
renaming.